Bowling score calculator
`BowlingProperties` is responsible for information related to a roll (ie. score, is strike, frame length, etc). `FrameScoreCalculator` is responsible for calculating the score of a single frame given a string of rolls. `Bowling` is responsible for parsing through the bowling frames in a given string and totalling the score of the frames.